fix: correct byte type in WriterIOAdapterExample error handling section
- Line 98: Changed write(65) to write(65.toByte) for type consistency
- Ensures Int literal is properly converted to Byte when writing to ByteArrayOutputStream

+++ streams-examples/src/main/scala/writer/WriterIOAdapterExample.scala
@@ -95,7 +95,7 @@ object WriterIOAdapterExample extends App {
   closedStream.close()
   val failingWriter = Writer.fromOutputStream(closedStream)
 
-  val writeResult = failingWriter.write(65) // 'A'
+  val writeResult = failingWriter.write(65.toByte) // 'A'
   println(s"Write to closed stream: $writeResult")
   println(s"Writer is closed: ${failingWriter.isClosed}")
 }
